Move the module-filter test into tests/test_fp8.py
Keeps it beside the other torchao coverage and drops the new
tests/quantization/torchao_integration directory this PR was creating. It sits
in its own class rather than in TestTorchAO because the layer swap needs no
accelerator and no huggingface suite, so it runs in-process instead of through
the launcher.
