+++ tests/test_fp8.py
@@ -285,6 +285,38 @@ class TestTorchAO(unittest.TestCase):
             run_command(command)
 
 
+@require_torchao
+class TestTorchAOModuleFilter(unittest.TestCase):
+    """The layer swap itself needs no accelerator, so this runs in-process rather than
+    through the launcher like the rest of the torchao tests above."""
+
+    def test_convert_model_to_fp8_ao_skips_the_first_and_last_linear_layers(self):
+        # convert_model_to_fp8_ao documents that it converts every nn.Linear "except the
+        # first and last", and find_first_last_linear_layers exists because quantizing
+        # those two destabilises training. The default module_filter_func was
+        # filter_first_and_last_linear_layers, which re-derives the first and last linear
+        # from the module it is handed. torchao hands it one candidate layer at a time, so
+        # it always found ("", "") and filtered nothing, and every linear including the
+        # first and last was converted.
+        from torchao.float8.float8_linear import Float8Linear
+
+        from accelerate.utils.ao import convert_model_to_fp8_ao
+
+        class ToyModel(torch.nn.Module):
+            def __init__(self):
+                super().__init__()
+                self.embed_proj = torch.nn.Linear(32, 64)
+                self.block = torch.nn.Sequential(torch.nn.Linear(64, 64), torch.nn.ReLU())
+                self.lm_head = torch.nn.Linear(64, 32)
+
+        model = ToyModel()
+        convert_model_to_fp8_ao(model)
+
+        assert not isinstance(model.embed_proj, Float8Linear)
+        assert not isinstance(model.lm_head, Float8Linear)
+        assert isinstance(model.block[0], Float8Linear)
+
+
 if __name__ == "__main__":
     # TE suite
     parser = argparse.ArgumentParser()
